+++ packages/ethereum-contracts/contracts/mocks/SuperTokenMock.t.sol
@@ -64,8 +64,14 @@ contract SuperTokenStorageLayoutTester is SuperToken {
         require (slot == 18 && offset == 0, "_operators changed location");
         // uses 4 slots
 
-        assembly { slot:= _reserve23.slot offset := _reserve23.offset }
-        require (slot == 23 && offset == 0, "_reserve23 changed location");
+        assembly { slot:= _nonces.slot offset := _nonces.offset }
+        require (slot == 22 && offset == 0, "_nonces changed location");
+
+        assembly { slot:= _yieldBackend.slot offset := _yieldBackend.offset }
+        require (slot == 23 && offset == 0, "_yieldBackend changed location");
+
+        assembly { slot:= _reserve24.slot offset := _reserve24.offset }
+        require (slot == 24 && offset == 0, "_reserve24 changed location");
 
         assembly { slot:= _reserve31.slot offset := _reserve31.offset }
         require (slot == 31 && offset == 0, "_reserve31 changed location");